Exercise set1 houndout

+++ src/academy/everyonecodes/java/week1/set1/exercise4/WateringMachine.java
@@ -3,14 +3,14 @@ package academy.everyonecodes.java.week1.set1.exercise4;
 public class WateringMachine {
 
     public static void main(String[] args) {
-        int numberOfPlants = 3;
+        double numberOfPlants = 3.0;
         int numberOfPlantsToBuy = 9;
-        int numberOfPlantsPerMachine = 4;
+        double numberOfPlantsPerMachine = 4.0;
         int numberOfOwnedMachines = 1;
         double costOfWaterMachine = 20.0;
-        double discount = 20 / 100;
+        double discount = 20.0 / 100.0;
         double discountedCost = costOfWaterMachine - (costOfWaterMachine * discount);
-        int totalNumberOfPlants = numberOfPlants + numberOfPlantsToBuy;
+        double totalNumberOfPlants = numberOfPlants + numberOfPlantsToBuy;
         double numberOfTotalMachinesToBeUsed = totalNumberOfPlants / numberOfPlantsPerMachine;
         double numberOfWateringMachinesToBuy = numberOfTotalMachinesToBeUsed - numberOfOwnedMachines;
         double finalCost = discountedCost * numberOfWateringMachinesToBuy;
